chore: create utils file
